requirements.txt
PySide6
python-mpv
PyGObject>=3.42.0
psutil
python-xlib
requests
Pillow
pydbus
